# oidc2-e2e-server/api/swagger.yaml
openapi: 3.0.3
info:
  title: Endpoint for establishing and maintaining an encrypted client-server session
  description: "This endpoint establishes an end-to-end encrypted client-server session\
    \ after authenticating the client with its Id Certification Token (Client_ICT).\
    \ After the successful exchange of DH parameters, the client can communicate with\
    \ the server, upload, download or delete files on the server with all communication\
    \ being encrypted on application layer."
  version: "0.1"
servers:
- url: http://rs.localhost
...
paths:
...
  /data:
    get:
      tags:
      - data
      summary: Download file from server
      description: Receive a GET file request and respond with the encrypted file.
        The file path in the GET request header is DH encrypted.
      responses:
        "200":
          description: |
            **OK**

            Contains the requested file DH encrypted in the body. Furthermore, it contains a new sessionToken and new server's DH public parameter in the header.
        "403":
          description: |
            **Forbidden**

            Possible reasons:
            - invalid/outdated sessionToken
            - invalid MACs
            - wrong encryption
        "500":
          description: |
            **Internal Server Error**

            Possible reasons:
            - couldn't find file
    post:
      tags:
      - data
      summary: Upload file to server
      description: Receive an encrypted file in the message body. Save it.
      operationId: postFile
      requestBody:
        description: Receive a JSON containing the encrypted file and a valid sessionToken.
        content:
          text/plain:
            schema:
              $ref: '#/components/schemas/UploadFile'
        required: true
      responses:
        "201":
          description: |
            **Created**

            Contains a DH encrypted location header to the new file (response to create), a status (response to delete) or a file (response to read). Furthermore, it contains a new sessionToken and new server's DH public parameter
        "403":
          description: |
            **Forbidden**

            Possible reasons:
            - invalid/outdated sessionToken
            - invalid MACs
            - wrong encryption
        "500":
          description: |
            **Internal Server Error**

            Possible reasons:
            - couldn't create file
    delete:
      tags:
      - data
      summary: Delete a file on the server
      description: Receive a delete request for a certain file on the server. File
        path is DH encrypted in the header.
      responses:
        "200":
          description: |
            **OK**

            The file was deleted successfully.
        "403":
          description: |
            **Forbidden**

            Possible reasons:
            - invalid/outdated sessionToken
            - invalid MACs
            - wrong encryption
        "500":
          description: |
            **Internal Server Error**

            Possible reasons:
            - couldn't find file
...
components:
  schemas:
...
    UploadFile:
      type: string
      description: |
        A HTTP POST request which contains an encrypted file in its body
      format: text/plain